improve: global Details Component refactoring
