+++ src/components/MusicStyleDetails/MusicStyleDetailsComponent.js
@@ -3,12 +3,16 @@
 import React, { Component } from "react";
 import "./styles/MusicStyleDetailsComponent.css";
 
-import ArtistsComponent from "./ArtistsComponent";
-import AnnecdoteComponent from "./AnecdoteComponent";
-import LinksComponent from "./LinksComponent";
-import ImpactComponent from "./ImpactComponent";
-import OriginComponent from "./OrigineComponent";
-import ThemeComponent from "./ThemeComponent";
+import ArtistsComponent from "./components/ArtistsComponent";
+import AnecdoteComponent from "./components/AnecdoteComponent";
+import LinksComponent from "./components/LinksComponent";
+import ImpactComponent from "./components/ImpactComponent";
+import OriginComponent from "./components/OrigineComponent";
+import ThemeComponent from "./components/ThemeComponent";
+
+import pointFreeUpperCase from "../../utils/pointFreeUpperCase";
+import { NavigationDetails } from "../NavigationBar/index";
+import { Context } from "../../App";
 
 // import Header Component
 import HeaderComponent from "../../utils/headerComponent";
@@ -31,51 +35,98 @@ export default class MusicStyleDetailsComponent extends Component<
 > {
   state = {};
 
+  renderContent = () => {
+    const {
+      params: { musicStyle, musicStyleDetail }
+    } = this.props;
+    switch (musicStyleDetail) {
+      case "artists":
+        return (
+          <ArtistsComponent
+            musicStyle={musicStyle}
+            musicStyleDetail={musicStyleDetail}
+          />
+        );
+      case "anecdotes":
+        return (
+          <AnecdoteComponent
+            musicStyle={musicStyle}
+            musicStyleDetail={musicStyleDetail}
+          />
+        );
+      case "links":
+        return (
+          <LinksComponent
+            musicStyle={musicStyle}
+            musicStyleDetail={musicStyleDetail}
+          />
+        );
+      case "blues":
+        switch (musicStyleDetail) {
+          case "impact":
+            return (
+              <ImpactComponent
+                musicStyle={musicStyle}
+                musicStyleDetail={musicStyleDetail}
+              />
+            );
+          case "origine":
+            return (
+              <OriginComponent
+                musicStyle={musicStyle}
+                musicStyleDetail={musicStyleDetail}
+              />
+            );
+          case "theme":
+            return (
+              <ThemeComponent
+                musicStyle={musicStyle}
+                musicStyleDetail={musicStyleDetail}
+              />
+            );
+          default:
+        }
+        break;
+      default:
+    }
+  };
+
   render() {
-    const { params } = this.props;
+    const {
+      params: { musicStyle, musicStyleDetail },
+      params
+    } = this.props;
 
     return (
       <div>
         <HeaderComponent params={params} />
-        {params.musicStyleDetail === "artists" && (
-          <ArtistsComponent
-            musicStyle={params.musicStyle}
-            musicStyleDetail={params.musicStyleDetail}
-          />
-        )}
-        {params.musicStyleDetail === "anecdotes" && (
-          <AnnecdoteComponent
-            musicStyle={params.musicStyle}
-            musicStyleDetail={params.musicStyleDetail}
-          />
-        )}
-        {params.musicStyleDetail === "links" && (
-          <LinksComponent
-            musicStyle={params.musicStyle}
-            musicStyleDetail={params.musicStyleDetail}
-          />
-        )}
-        {params.musicStyle === "blues" &&
-        params.musicStyleDetail === "impact" ? (
-          <ImpactComponent
-            musicStyle={params.musicStyle}
-            musicStyleDetail={params.musicStyleDetail}
-          />
-        ) : null}
-        {params.musicStyle === "blues" &&
-        params.musicStyleDetail === "origine" ? (
-          <OriginComponent
-            musicStyle={params.musicStyle}
-            musicStyleDetail={params.musicStyleDetail}
-          />
-        ) : null}
-        {params.musicStyle === "blues" &&
-        params.musicStyleDetail === "themes" ? (
-          <ThemeComponent
-            musicStyle={params.musicStyle}
-            musicStyleDetail={params.musicStyleDetail}
-          />
-        ) : null}
+        <Context.Consumer>
+          {({ MUSIC_DETAILS, BLUES_DETAILS }) => (
+            <div id="wrap">
+              <div className="flex">
+                <h1>
+                  <span>{pointFreeUpperCase(musicStyleDetail)}</span>
+                </h1>
+                {this.renderContent()}
+              </div>
+              <ul className="navDetails">
+                {musicStyle === "blues" ? (
+                  <NavigationDetails
+                    arrayElement={BLUES_DETAILS}
+                    musicStyle={musicStyle}
+                    currentDetail={musicStyleDetail}
+                  />
+                ) : (
+                  <NavigationDetails
+                    arrayElement={MUSIC_DETAILS}
+                    musicStyle={musicStyle}
+                    currentDetail={musicStyleDetail}
+                  />
+                )}
+              </ul>
+            </div>
+          )}
+        </Context.Consumer>
       </div>
     );
   }
